Backend/requirements.txt
# Core Django and DRF
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1

# IMPORTANT: setuptools 68.x includes pkg_resources which is required by drf-yasg
# Newer versions (69+) have deprecated pkg_resources causing import errors
setuptools==68.0.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication and Security
django-allauth==0.61.1
dj-rest-auth==5.0.2
PyJWT==2.8.0
django-oauth-toolkit==2.3.0
social-auth-app-django==5.4.0

# OTP and Phone Verification
twilio==9.0.4
pyotp==2.9.0
qrcode==8.0

# Payment
stripe==8.2.0

# File Handling
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.51

# Environment Variables
python-dotenv==1.0.1

# API Documentation
# drf-yasg requires pkg_resources from setuptools<=68
drf-yasg==1.21.7
drf-spectacular==0.27.1
uritemplate==4.1.1
inflection==0.5.1
pyyaml==6.0.2

# Utilities
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
whitenoise==6.6.0
python-dateutil==2.9.0

# Development Tools
django-debug-toolbar==4.3.0
django-extensions==3.2.3

# Production Server
gunicorn==21.2.0
uvicorn==0.27.1

# Testing
pytest==8.0.2
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==24.0.0

# Additional Dependencies
pytz==2024.1
sqlparse==0.5.0
asgiref==3.8.1
